[ASTGen] Move some code to MiscBridging.cpp
Also this fixes a weird 'undefined symbol' error when linking with gold.

rdar://137326126
